# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Portfolio API",
        "description": "Laravel12で構築したWeb APIのドキュメントです",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "ユーザー登録を行う",
                "operationId": "20b4178f980c05a589875bee66120ac0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "ユーザー名",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Taro Yamada"
                                    },
                                    "email": {
                                        "description": "メールアドレス",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "パスワード",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ユーザー登録成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "バリデーションエラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "ユーザーアカウントを削除する",
                "operationId": "539b5fc3022cbcb3757d9a8b9becb4c8",
                "responses": {
                    "201": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ユーザーが見つかりません",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "ユーザーログインを行う",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "メールアドレス",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "パスワード",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ログイン成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "バリデーションエラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "ユーザーログアウトを行う",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "ログアウト成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Unauthenticated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "Unauthorized": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized."
                    }
                },
                "type": "object"
            },
...
            "NotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required. (and 1 more error)"
                    },
                    "errors": {
                        "description": "各フィールドごとのエラーメッセージ",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The title field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Taro Yamada"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctumで発行されたBearerトークンを入力してください",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}